+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   resource agent metadata ([rhbz#1665404])
 - Command `pcs resource bundle reset` for a bundle configuration reseting
   ([rhbz#1657166])
+- `pcs cluster setup` now checks if nodes' addresses match value of `ip_version`
+  ([rhbz#1667053])
 
 ### Fixed
 - Corosync config file parser updated and made more strict to match changes in
@@ -37,6 +39,9 @@
 - Lower load caused by periodical config files syncing in pcsd by making it
   sync less frequently ([rhbz#1676957])
 - Improve logging of periodical config files syncing in pcsd
+- Knet link option `ip_version` has been removed, it was never supported by
+  corosync. Transport option `ip_version` is still in place. ([rhbz#1674005])
+- Several bugs in linklist validation in `pcs cluster setup` ([rhbz#1667090])
 
 ### Deprecated
 - Command `pcs resource show`, removed in pcs-0.10.1, has been readded as
@@ -53,8 +58,11 @@
 [rhbz#1664057]: https://bugzilla.redhat.com/show_bug.cgi?id=1664057
 [rhbz#1665404]: https://bugzilla.redhat.com/show_bug.cgi?id=1665404
 [rhbz#1667040]: https://bugzilla.redhat.com/show_bug.cgi?id=1667040
+[rhbz#1667053]: https://bugzilla.redhat.com/show_bug.cgi?id=1667053
+[rhbz#1667090]: https://bugzilla.redhat.com/show_bug.cgi?id=1667090
 [rhbz#1668223]: https://bugzilla.redhat.com/show_bug.cgi?id=1668223
 [rhbz#1668422]: https://bugzilla.redhat.com/show_bug.cgi?id=1668422
+[rhbz#1674005]: https://bugzilla.redhat.com/show_bug.cgi?id=1674005
 [rhbz#1676945]: https://bugzilla.redhat.com/show_bug.cgi?id=1676945
 [rhbz#1676957]: https://bugzilla.redhat.com/show_bug.cgi?id=1676957